+++ lib/qiita/elasticsearch/query_builder.rb
@@ -1,3 +1,4 @@
+require "qiita/elasticsearch/errors"
 require "qiita/elasticsearch/nodes/null_node"
 require "qiita/elasticsearch/nodes/or_separatable_node"
 require "qiita/elasticsearch/tokenizer"
@@ -5,17 +6,21 @@ require "qiita/elasticsearch/tokenizer"
 module Qiita
   module Elasticsearch
     class QueryBuilder
+      # @param [Array<String>, nil] date_fields
       # @param [Array<String>, nil] downcased_fields
       # @param [Array<String>, nil] filterable_fields
       # @param [Array<String>, nil] hierarchal_fields
+      # @param [Array<String>, nil] int_fields
       # @param [Array<String>, nil] matchable_fields
-      # @param [Array<String>, nil] range_fields
-      def initialize(downcased_fields: nil, hierarchal_fields: nil, filterable_fields: nil, matchable_fields: nil, range_fields: nil)
+      # @param [String, nil] time_zone
+      def initialize(date_fields: nil, downcased_fields: nil, hierarchal_fields: nil, filterable_fields: nil, int_fields: nil, matchable_fields: nil, time_zone: nil)
+        @date_fields = date_fields
         @downcased_fields = downcased_fields
-        @hierarchal_fields = hierarchal_fields
         @filterable_fields = filterable_fields
+        @hierarchal_fields = hierarchal_fields
+        @int_fields = int_fields
         @matchable_fields = matchable_fields
-        @range_fields = range_fields
+        @time_zone = time_zone
       end
 
       # @param [String] query_string Raw query string
@@ -27,17 +32,21 @@ module Qiita
         else
           Nodes::OrSeparatableNode.new(tokens).to_hash
         end
+      rescue Error
+        Nodes::NullNode.new.to_hash
       end
 
       private
 
       def tokenizer
         @tokenizer ||= Tokenizer.new(
+          date_fields: @date_fields,
           downcased_fields: @downcased_fields,
-          hierarchal_fields: @hierarchal_fields,
           filterable_fields: @filterable_fields,
+          hierarchal_fields: @hierarchal_fields,
+          int_fields: @int_fields,
           matchable_fields: @matchable_fields,
-          range_fields: @range_fields,
+          time_zone: @time_zone,
         )
       end
     end