+++ lib/qiita/elasticsearch/tokenizer.rb
@@ -1,15 +1,17 @@
+require "qiita/elasticsearch/date_token"
 require "qiita/elasticsearch/filterable_token"
 require "qiita/elasticsearch/hierarchal_token"
 require "qiita/elasticsearch/matchable_token"
-require "qiita/elasticsearch/range_token"
+require "qiita/elasticsearch/int_token"
 
 module Qiita
   module Elasticsearch
     class Tokenizer
+      DEFAULT_DATE_FIELDS = []
       DEFAULT_DOWNCASED_FIELDS = []
       DEFAULT_FILTERABLE_FIELDS = []
       DEFAULT_HIERARCHAL_FIELDS = []
-      DEFAULT_RANGE_FIELDS = []
+      DEFAULT_INT_FIELDS = []
 
       TOKEN_PATTERN = /
         (?<token_string>
@@ -23,17 +25,21 @@ module Qiita
         )
       /x
 
+      # @param [Array<String>, nil] date_fields
       # @param [Array<String>, nil] downcased_fields
       # @param [Array<String>, nil] filterable_fields
       # @param [Array<String>, nil] hierarchal_fields
+      # @param [Array<String>, nil] int_fields
       # @param [Array<String>, nil] matchable_fields
-      # @param [Array<String>, nil] range_fields
-      def initialize(downcased_fields: nil, filterable_fields: nil, hierarchal_fields: nil, matchable_fields: nil, range_fields: nil)
+      # @param [String, nil] time_zone
+      def initialize(date_fields: nil, downcased_fields: nil, filterable_fields: nil, hierarchal_fields: nil, int_fields: nil, matchable_fields: nil, time_zone: nil)
+        @date_fields = date_fields
         @downcased_fields = downcased_fields
         @filterable_fields = filterable_fields
         @hierarchal_fields = hierarchal_fields
+        @int_fields = int_fields
         @matchable_fields = matchable_fields
-        @range_fields = range_fields
+        @time_zone = time_zone
       end
 
       # @param [String] query_string Raw query string
@@ -54,11 +60,16 @@ module Qiita
             token_string: token_string,
           )
           token.matchable_fields = @matchable_fields if token.is_a?(MatchableToken)
+          token.time_zone = @time_zone if token.is_a?(DateToken)
           token
         end
       end
 
       private
+
+      def date_fields
+        @date_fields || DEFAULT_DATE_FIELDS
+      end
 
       def downcased_fields
         @downcased_fields || DEFAULT_DOWNCASED_FIELDS
@@ -72,14 +83,16 @@ module Qiita
         @hierarchal_fields || DEFAULT_HIERARCHAL_FIELDS
       end
 
-      def range_fields
-        @range_fields || DEFAULT_RANGE_FIELDS
+      def int_fields
+        @int_fields || DEFAULT_INT_FIELDS
       end
 
       def token_class(field_name)
         case
-        when range_fields.include?(field_name)
-          RangeToken
+        when date_fields.include?(field_name)
+          DateToken
+        when int_fields.include?(field_name)
+          IntToken
         when hierarchal_fields.include?(field_name)
           HierarchalToken
         when filterable_fields.include?(field_name)